requirements.txt
# Core dependencies
anthropic==0.39.0
pandas==2.2.0
numpy==1.26.3

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.11.0

# Text processing
python-docx==1.1.0
openpyxl==3.1.2

# Progress tracking
tqdm==4.66.1

# Environment variables
python-dotenv==1.0.0

# Jupyter (optional, for exploration)
jupyter==1.0.0
ipykernel==6.28.0

# Visualization (optional)
matplotlib==3.8.2
seaborn==0.13.1

# Embeddings (for bonus task)
sentence-transformers==2.3.1
faiss-cpu==1.7.4

pip install datasets